Match Julia project structure where: - module name includes all exports, imports, and includes all   sibling source files. - module file is upper first CamelCase - include files are lower first camelCase.
