+++ src/Transit.jl
@@ -1,0 +1,9 @@
+module Transit
+  import JSON
+
+  export Encoder, encode
+
+  include("encoder.jl")
+  include("emitter.jl")
+  include("rollingCache.jl")
+end
