+++ src/encoder.jl
@@ -1,25 +1,20 @@
-module Transit
-  export Encoder, encode
+type Encoder
 
-  type Encoder
-    
-  end
-
-  function encode(e::Encoder, s::AbstractString)
-    println("string: ", s)
-  end
-
-  function encode(e::Encoder, i::Integer)
-    println("number:", i)
-  end
-
-  function encode(e::Encoder, i::Rational)
-    println("rational:", i)
-  end
-
-  function encode(e::Encoder, a::Array)
-    println("array:  ", a)
-    println("array:  ", a[1])
-  end
 end
 
+function encode(e::Encoder, s::AbstractString)
+  println("string: ", s)
+end
+
+function encode(e::Encoder, i::Integer)
+  println("number:", i)
+end
+
+function encode(e::Encoder, i::Rational)
+  println("rational:", i)
+end
+
+function encode(e::Encoder, a::Array)
+  println("array:  ", a)
+  println("array:  ", a[1])
+end
